fixed implicit type casting (warning)

+++ cpp/src/pieces/pawn.cpp
@@ -56,7 +56,7 @@ std::vector<Move> pawn_moves(const Board& b, const Coords xy) {
                 ret.push_back(Move{
                     xy.to_index(),
                     new_xy.to_index(),
-                    .promoting_to = my_colour | piece
+                    .promoting_to = (int8_t) (my_colour | piece)
                 });
     }
